requirements.txt
requests
yadisk[sync-defaults]
python-dotenv
python-telegram-bot[job-queue]